+++ spec/integration/resource/api600/synergy/volume/delete_spec.rb
@@ -1,8 +1,8 @@
 require 'spec_helper'
 
-klass = OneviewSDK::API500::Synergy::Volume
+klass = OneviewSDK::API600::Synergy::Volume
 RSpec.describe klass, integration: true, type: DELETE, sequence: rseq(klass) do
-  let(:current_client) { $client_500_synergy }
-  include_examples 'VolumeDeleteExample', 'integration api500 context'
-  include_examples 'VolumeDeleteExample API500', 'integration api500 context'
+  let(:current_client) { $client_600_synergy }
+  include_examples 'VolumeDeleteExample', 'integration api600 context'
+  include_examples 'VolumeDeleteExample API600', 'integration api600 context'
 end